优化入库方式

+++ src/main/com/handu/scada/quartz/job/lp/DeviceInfoChangeJob.java
@@ -8,8 +8,8 @@ import main.com.handu.scada.db.bean.common.DeviceDtuCacheResult;
 import main.com.handu.scada.db.mapper.DeviceInfochangesMapper;
 import main.com.handu.scada.db.mapper.common.CommonMapper;
 import main.com.handu.scada.db.utils.MyBatisUtil;
-import main.com.handu.scada.quartz.job.CommonJob;
 import main.com.handu.scada.quartz.job.BaseJob;
+import main.com.handu.scada.quartz.job.CommonJob;
 import main.com.handu.scada.utils.DateUtils;
 import main.com.handu.scada.utils.LogUtils;
 import org.apache.ibatis.session.SqlSession;
@@ -18,9 +18,8 @@ import org.quartz.JobExecutionException;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Vector;
+import java.util.concurrent.ConcurrentHashMap;
 
-import static main.com.handu.scada.cache.MyCacheManager.DEVICE_DTU_INFO;
 
 /**
  * Created by 柳梦 on 2018/01/03.
@@ -36,7 +35,7 @@ public class DeviceInfoChangeJob extends CommonJob implements BaseJob {
 
     @Override
     public String cronExpression() {
-        return "0/30 * * * * ?";
+        return "45 * * * * ?";
     }
 
     @Override
@@ -52,13 +51,12 @@ public class DeviceInfoChangeJob extends CommonJob implements BaseJob {
                     //初始化设备相关
                     List<DeviceDtuCacheResult> list = commonMapper.selectDeviceDtuCacheResult(Arrays.asList(Config.getDtuPorts().split(",")));
                     if (list != null) {
-                        Vector<DeviceDtuCacheResult> results = MyCacheManager.getInstance().getDataByKey(DEVICE_DTU_INFO);
-                        synchronized (results) {
-                            results.clear();
-                            results.addAll(list);
+                        ConcurrentHashMap<String, DeviceDtuCacheResult> cacheResults = MyCacheManager.getInstance().getDeviceDtuCache();
+                        if (cacheResults != null) {
+                            MyCacheManager.getInstance().putDeviceDtuCacheResult(list);
+                            LogUtils.info("init device cache result count " + list.size(), true);
+                            infochanges.forEach(deviceInfochanges -> mapper.deleteByPrimaryKey(deviceInfochanges.getOid()));
                         }
-                        LogUtils.info("init device cache result count " + list.size(), true);
-                        infochanges.forEach(deviceInfochanges -> mapper.deleteByPrimaryKey(deviceInfochanges.getOid()));
                     }
                 }
                 if (sqlSession != null) sqlSession.commit(true);
